made more tables in firebase with connections to code
made more tables and made a way for the data base to populate the page as well as updating the table such as deleting and adding. the page will always constantly update these changes even if you sign out and come back, the changes will be updated and it is separate for every user.